requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.7
cron-descriptor==1.4.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.4
django-admin-rangefilter==0.10.0
django-auditlog==2.3.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-debug-toolbar==4.1.0
django-filter==23.2
django-import-export==3.2.0
django-js-asset==2.1.0
django-mptt==0.14.0
django-timezone-field==6.0
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-bulk==0.2.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
docstring-to-markdown==0.11
drf-yasg==1.21.7
et-xmlfile==1.1.0
exceptiongroup==1.1.2
Faker==19.2.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.1
MarkupPy==1.14
MarkupSafe==2.1.3
mysqlclient==2.0.3
numpy==1.25.1
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycodestyle==2.10.0
pydocstyle==6.2.3
pyflakes==3.0.1
PyJWT==1.7.1
pytest==7.4.0
pytest-django==4.5.2
python-crontab==3.0.0
python-dateutil==2.8.2
pytoolconfig==1.2.5
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.4
tablib==3.5.0
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
uWSGI==2.0.21
vine==5.0.0
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
